refactor: . Move createSetStatusCommands() to CreateSetStatusLineTransformer.ts
This is so that it can be tested.
